Step 4: front end update
