+++ backend/retrieval.py
@@ -8,40 +8,12 @@ torch (a heavy dependency) is imported lazily inside ``retrieve`` since it is
 only needed when actually embedding a query.
 """
 
-from pydantic import BaseModel
 from sentence_transformers import CrossEncoder, SentenceTransformer
 from sqlalchemy import select
 from sqlalchemy.orm import Session
 
+from backend.schemas import RetrievedChunk
 from db.models import ChunkDim384
-
-
-class RetrievedChunk(BaseModel):
-    """A single document chunk returned by retrieval/reranking.
-
-    Identity is the ``(source, start_index)`` pair. ``chunk_id`` is the DB
-    primary key, carried along for eval joins/debugging but not relied upon.
-    """
-
-    source: str
-    start_index: int
-    text: str
-    retrieval_score: float | None = None
-    rerank_score: float | None = None
-    chunk_id: int | None = None
-
-
-class RetrieveRequest(BaseModel):
-    query: str
-    vector_store_config_id: int
-    top_k: int
-    source: str | None = None
-
-
-class RerankRequest(BaseModel):
-    query: str
-    chunks: list[RetrievedChunk]
-    top_k: int
 
 
 def retrieve(
