+++ frontend/client.py
@@ -5,7 +5,7 @@ import jinja2
 from dotenv import load_dotenv
 from openai import OpenAI
 
-from backend.retrieval import RetrievedChunk
+from backend.schemas import RetrievedChunk
 
 load_dotenv()
